+++ src/systems/mp-units-systems.cpp
@@ -36,6 +36,7 @@ import std;
 
 #include <mp-units/systems/angular.h>
 #include <mp-units/systems/cgs.h>
+#include <mp-units/systems/codata.h>
 #include <mp-units/systems/hep.h>
 #include <mp-units/systems/iau.h>
 #include <mp-units/systems/iec.h>